fix: fix package
