+++ packages/checkbox/package.json
@@ -30,7 +30,7 @@
   "repository": {
     "type": "git",
     "url": "git+https://github.com/illa-family/illa-design.git",
-    "directory": "packages/radio"
+    "directory": "packages/checkbox"
   },
   "bugs": {
     "url": "https://github.com/illa-family/illa-design/issues"
