+++ packages/empty/package.json
@@ -30,7 +30,7 @@
   "repository": {
     "type": "git",
     "url": "git+https://github.com/illa-family/illa-design.git",
-    "directory": "packages/icon"
+    "directory": "packages/empty"
   },
   "bugs": {
     "url": "https://github.com/illa-family/illa-design/issues"
