+++ packages/input/src/style.tsx
@@ -1,9 +1,8 @@
-import { InputSize } from "./interface"
+import { InputSize, StateValue } from "./interface"
 
-const chroma = require("chroma-js")
+import chroma from "chroma-js"
 import { css, SerializedStyles } from "@emotion/react"
 import { globalColor, illaPrefix } from "@illa-design/theme"
-import { StateValue } from "./input"
 
 export const inputFillStyle = css`
   background-color: ${globalColor(`--${illaPrefix}-gray-09`)};
