+++ packages/input/src/textarea-style.tsx
@@ -1,4 +1,4 @@
-const chroma = require("chroma-js")
+import chroma from "chroma-js"
 import { css, SerializedStyles } from "@emotion/react"
 import { globalColor, illaPrefix } from "@illa-design/theme"
 import {
@@ -11,7 +11,7 @@ import {
   inputFillStyle,
   inputOutlineStyle,
 } from "./style"
-import { StateValue } from "./input"
+import { StateValue } from "./interface"
 
 function applyStatus(stateValue: StateValue) {
   let mainStyle: SerializedStyles
