+++ packages/react/package.json
@@ -55,6 +55,7 @@
     "@illa-design/tooltip": "*",
     "@illa-design/popover": "*",
     "@illa-design/link": "*",
+    "@illa-design/config-provider": "*",
     "@illa-design/progress": "*",
     "@illa-design/empty": "*",
     "@illa-design/badge": "*",
@@ -63,6 +64,7 @@
     "@illa-design/list": "*",
     "@illa-design/switch": "*",
     "@illa-design/alert": "*",
+    "@illa-design/input-tag": "*",
     "@illa-design/grid": "*",
     "@illa-design/pagination": "*",
     "@illa-design/upload": "*",
