+++ packages/react/src/index.ts
@@ -1,9 +1,10 @@
+export * from "@illa-design/system"
+export * from "@illa-design/theme"
+export * from "@illa-design/config-provider"
 export * from "@illa-design/avatar"
 export * from "@illa-design/icon"
 export * from "@illa-design/image"
-export * from "@illa-design/system"
 export * from "@illa-design/tag"
-export * from "@illa-design/theme"
 export * from "@illa-design/divider"
 export * from "@illa-design/space"
 export * from "@illa-design/button"
@@ -14,9 +15,17 @@ export * from "@illa-design/tooltip"
 export * from "@illa-design/popover"
 export * from "@illa-design/input"
 export * from "@illa-design/link"
-export * from "@illa-design/config-provider"
 export * from "@illa-design/progress"
 export * from "@illa-design/empty"
+export * from "@illa-design/badge"
+export * from "@illa-design/statistic"
 export * from "@illa-design/checkbox"
 export * from "@illa-design/list"
+export * from "@illa-design/switch"
+export * from "@illa-design/alert"
+export * from "@illa-design/input-tag"
 export * from "@illa-design/grid"
+export * from "@illa-design/pagination"
+export * from "@illa-design/upload"
+export * from "@illa-design/spin"
+export * from "@illa-design/notification"
